requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-cors-headers==4.0.0
django-model-utils==4.3.1
django-tables2==2.5.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
xlwt==1.3.0
